refactored sample name
