+++ generate-offline-map-using-android-jetpack-workmanager/README.metadata.json
@@ -1,7 +1,7 @@
 {
     "category": "Edit and Manage Data",
     "description": "Take a web map offline as a persistent background task",
-    "formal_name": "GenerateOfflineMapUsingWorkmanager",
+    "formal_name": "GenerateOfflineMapUsingAndroidJetpackWorkmanager",
     "ignore": false,
     "images": [
         "generate-offline-map.png"
@@ -31,10 +31,10 @@
         "Portal"
     ],
     "snippets": [
-        "src/main/java/com/esri/arcgismaps/sample/generateofflinemapusingworkmanager/MainActivity.kt",
-        "src/main/java/com/esri/arcgismaps/sample/generateofflinemapusingworkmanager/OfflineJobWorker.kt",
-        "src/main/java/com/esri/arcgismaps/sample/generateofflinemapusingworkmanager/WorkerNotification.kt",
-        "src/main/java/com/esri/arcgismaps/sample/generateofflinemapusingworkmanager/NotificationActionReceiver.kt"
+        "src/main/java/com/esri/arcgismaps/sample/generateofflinemapusingandroidjetpackworkmanager/MainActivity.kt",
+        "src/main/java/com/esri/arcgismaps/sample/generateofflinemapusingandroidjetpackworkmanager/OfflineJobWorker.kt",
+        "src/main/java/com/esri/arcgismaps/sample/generateofflinemapusingandroidjetpackworkmanager/WorkerNotification.kt",
+        "src/main/java/com/esri/arcgismaps/sample/generateofflinemapusingandroidjetpackworkmanager/NotificationActionReceiver.kt"
     ],
-    "title": "Generate offline map using workmanager"
+    "title": "Generate offline map using Android Jetpack WorkManager"
 }
