+++ generate-offline-map-using-android-jetpack-workmanager/build.gradle
@@ -5,7 +5,7 @@ android {
     compileSdkVersion rootProject.ext.compileSdkVersion
 
     defaultConfig {
-        applicationId "com.esri.arcgismaps.sample.generateofflinemapusingworkmanager"
+        applicationId "com.esri.arcgismaps.sample.generateofflinemapusingandroidjetpackworkmanager"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode rootProject.ext.versionCode
@@ -24,7 +24,7 @@ android {
         dataBinding true
     }
 
-    namespace 'com.esri.arcgismaps.sample.generateofflinemapusingworkmanager'
+    namespace 'com.esri.arcgismaps.sample.generateofflinemapusingandroidjetpackworkmanager'
 }
 
 dependencies {
