+++ generate-offline-map-using-android-jetpack-workmanager/src/main/java/com/esri/arcgismaps/sample/generateofflinemapusingandroidjetpackworkmanager/WorkerNotification.kt
@@ -15,7 +15,7 @@
  *
  */
 
-package com.esri.arcgismaps.sample.generateofflinemapusingworkmanager
+package com.esri.arcgismaps.sample.generateofflinemapusingandroidjetpackworkmanager
 
 import android.app.Notification
 import android.app.NotificationChannel
