+++ generate-offline-map-using-android-jetpack-workmanager/src/main/res/drawable-v24/ic_launcher_foreground.xml
