+++ generate-offline-map-using-android-jetpack-workmanager/src/main/res/layout/activity_main.xml
