+++ generate-offline-map-using-android-jetpack-workmanager/src/main/res/layout/generate_offline_map_dialog_layout.xml
