+++ generate-offline-map-using-android-jetpack-workmanager/src/main/res/mipmap-anydpi-v26/ic_launcher.xml
