+++ generate-offline-map-using-android-jetpack-workmanager/src/main/res/mipmap-hdpi/ic_launcher.png
