+++ generate-offline-map-using-android-jetpack-workmanager/src/main/res/mipmap-mdpi/ic_launcher.png
