+++ generate-offline-map-using-android-jetpack-workmanager/src/main/res/mipmap-xhdpi/ic_launcher.png
