+++ generate-offline-map-using-android-jetpack-workmanager/src/main/res/mipmap-xxhdpi/ic_launcher.png
