+++ generate-offline-map-using-android-jetpack-workmanager/src/main/res/mipmap-xxxhdpi/ic_launcher.png
