+++ generate-offline-map-using-android-jetpack-workmanager/src/main/res/values/colors.xml
