+++ generate-offline-map-using-android-jetpack-workmanager/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">Generate offline map using workmanager</string>
+    <string name="app_name">Generate offline map using android jetpack workmanager</string>
     <string name="portal_url">https://www.arcgis.com</string>
     <string name="item_id">acc027394bc84c2fb04d1ed317aac674</string>
     <string name="dialog_title">Generating offline map..</string>
